Day 19!
This was an interesting one because at the beginning I didn't really know if it would work.  My approach was to collect all the distances between the beacons and look for overlapping sets of distances that were at least size 66 (11+10+9+...)

This got me a correct answer for part 1, but an incorrect one for part 2.  This turned out to be because sometimes there were spurious distances that matched even when the beacons in question were not the same beacons.  It was easy to work out which ones were problematic because it would be the only one of the set that would match. (i.e. if one beacon among 12 actually matched, it would have 11 matches in the set with its distance to every other one).