Day 9!
I was happy that I wrote the function tailCatchUp for part 1 just because it seemed like a good thing to write a function for, and then it turned out to be critical for part 2.  Originally had a dumb bug in in for part 2 where I had sequential 'if' statements instead of 'elsif's.